+++ kite-core/src/main/kotlin/com/tang/kite/datasource/unpooled/UnpooledDataSourceFactory.kt
@@ -13,12 +13,8 @@ class UnpooledDataSourceFactory(private val properties: UnpooledProperties) : Da
     }
 
     override fun getDataSource(): DataSource {
-        if (properties.driver == null) {
-            throw IllegalArgumentException("driver is required")
-        }
-        if (properties.url == null) {
-            throw IllegalArgumentException("url is required")
-        }
+        requireNotNull(properties.driver) { "driver is required" }
+        requireNotNull(properties.url) { "url is required" }
         return UnpooledDataSource(properties)
     }